src/main/java/com/codessquad/qna/service/UserService.java
package com.codessquad.qna.service;

import com.codessquad.qna.domain.User;
import com.codessquad.qna.repository.UserRepository;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.Optional;

@Service
@Transactional(readOnly = true)
public class UserService {

    private UserRepository userRepository;
    Logger logger = LoggerFactory.getLogger(UserService.class);

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Transactional
    public Long join(User user) {
        User DBUser = userRepository.save(user);
        return DBUser.getId();
    }

    public Optional<User> findById(Long userId) {
        return userRepository.findById(userId);
    }

    public List<User> findAll() {
        return userRepository.findAll();
    }

    @Transactional
    public void update(User user, String newPassword) {
        User getUser = userRepository.findById(user.getId()).orElseThrow(IllegalArgumentException::new);
        if (getUser.checkPassword(user)) {
            getUser.updateUserInfo(user, newPassword);
        }
    }

}
